connect -> sky renaming
