+++ apps/cli/src/__tests__/__sky__/fetchFromStately.sky.ts
