+++ apps/cli/src/__tests__/__sky__/tsconfig.json
