+++ apps/cli/tsconfig.json
@@ -12,5 +12,5 @@
     "types": ["vite/client", "jest"]
   },
   "include": ["src"],
-  "exclude": ["src/__tests__/__examples__", "src/__tests__/__connect__"]
+  "exclude": ["src/__tests__/__examples__", "src/__tests__/__sky__"]
 }
