+++ packages/machine-extractor/src/getMachineNodesFromFile.ts
@@ -1,8 +1,8 @@
 import { parse } from '@babel/parser';
 import traverse from '@babel/traverse';
 import * as t from '@babel/types';
-import { ALLOWED_SKY_CONFIG_CALL_EXPRESSION_NAMES } from './connect/skyConfigUtils';
 import { ALLOWED_CALL_EXPRESSION_NAMES } from './machineCallExpression';
+import { ALLOWED_SKY_CONFIG_CALL_EXPRESSION_NAMES } from './sky/skyConfigUtils';
 
 export const getMachineNodesFromFile = (fileContent: string) => {
   const file = parse(fileContent, {
