+++ packages/shared/src/index.ts
@@ -1,7 +1,3 @@
-// TODO: everything here from `/connect` should come from the `@stately/connect` package
-export * from './connect/fetchFromStately';
-export * from './connect/fromStately';
-export * from './connect/skyTypes';
 export * from './createIntrospectableMachine';
 export * from './filterOutIgnoredMachines';
 export * from './forEachAction';
@@ -18,5 +14,9 @@ export * from './introspectMachine';
 export * from './isCursorInPosition';
 export * from './processFileEdits';
 export * from './resolveUriToFilePrefix';
+// TODO: everything here from `/sky` should come from the `@stately/sky` package
+export * from './sky/fetchFromStately';
+export * from './sky/fromStately';
+export * from './sky/skyTypes';
 export * from './types';
 export * from './writeToFetchedMachineFile';
