+++ packages/shared/src/sky/skyTypes.ts
@@ -1,6 +1,6 @@
 import { AnyStateMachine } from 'xstate';
 
-// TODO: use this from the Connect SDK package once that's released
+// TODO: consume these from the Sky SDK package once that's released
 export interface SkyConfig {
   config: any;
   actor: { id: string };
